Update majority of logic
was mostly saving progress locally without committing as I haven't
decided if I want to keep my logic this way. For now everything is
working besides . but I need to handle a way to limit the length
of digits or their display in a way that feels natural for users.